requirements.txt
langchain==0.1.12
langchain-openai
langchain-google-genai
fastapi==0.109.2
uvicorn==0.27.0.post1
llama-index
typer==0.9.0
streamlit==1.31.1
apscheduler==3.10.4
loguru==0.7.2
python-dotenv==1.0.1
sqlalchemy==2.0.27
sqlite-utils==3.35